@@ -1,9 +1,9 @@
-const cloudinary = require("cloudinary").v2;
+import cloudinary from "cloudinary";
 
 const cloudinaryConnector = () => {
     try {
         // Configure Cloudinary with environment variables
-        cloudinary.config({
+        cloudinary.v2.config({
             cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
             api_key: process.env.CLOUDINARY_API_KEY,
             api_secret: process.env.CLOUDINARY_API_SECRET,
@@ -21,4 +21,5 @@ const cloudinaryConnector = () => {
         );
     }
 };
-module.exports = cloudinaryConnector;
+
+export default cloudinaryConnector;
