@@ -1,167 +1,151 @@
 // This file includes the following controllers:
-//
 // 1. resetPasswordTokenMail
 // 2. resetPassword
 // 3. changePassword
 
-const bcrypt = require("bcrypt");
-const User = require("../models/User");
-const mailSender = require("../utils/mailSender");
-const crypto = require("crypto"); // Import crypto module for generating reset tokens
-const passwordResetLinkTemplate = require("../mail/templates/passwordResetLinkTemplate");
-const passwordUpdatedTemplate = require("../mail/templates/passwordUpdatedTemplate");
+import User from "../models/User.js";
+// Import necessary modules using ESM
+import bcrypt from "bcrypt";
+import crypto from "crypto"; // For generating secure reset tokens
+import dotenv from "dotenv"; // Load environment variables
+import mailSender from "../utils/mailSender.js";
+import passwordResetLinkTemplate from "../mail/templates/passwordResetLinkTemplate.js";
+import passwordUpdatedTemplate from "../mail/templates/passwordUpdatedTemplate.js";
+
+dotenv.config(); // Initialize environment variables
+
 // Function to send password reset link to the user's email
-exports.resetPasswordTokenMail = async (req, res) => {
+export const resetPasswordTokenMail = async (req, res) => {
     try {
         const { email } = req.body;
 
-        // Check if email is provided
+        // Validate input
         if (!email) {
             return res
                 .status(400)
-                .json({ success: false, message: "Email is required" });
+                .json({ success: false, message: "Email is required." });
         }
 
-        // Check if the user exists
+        // Check if user exists in the database
         const user = await User.findOne({ email });
         if (!user) {
             return res.status(404).json({
                 success: false,
-                message: "No account found with the provided email",
+                message: "No account found with this email.",
             });
         }
 
-        // Generate a unique reset password token
+        // Generate a unique reset token with an expiration time
         const resetPasswordToken = crypto.randomUUID();
-        const resetPasswordTokenExpires = Date.now() + 15 * 60 * 1000; // Token expires in 15 minutes
+        const resetPasswordTokenExpires = Date.now() + 15 * 60 * 1000; // Token valid for 15 minutes
 
-        // Update the user with the reset password token and expiration time
-        await User.findOneAndUpdate(
-            { email },
-            {
-                resetPasswordToken,
-                resetPasswordTokenExpires,
-            }
-        );
+        // Save the token to the user document
+        user.resetPasswordToken = resetPasswordToken;
+        user.resetPasswordTokenExpires = resetPasswordTokenExpires;
+        await user.save();
 
-        // Create the password reset URL
+        // Construct password reset URL
         const resetPasswordUrl = `${process.env.FRONTEND_URL}/reset-password/${resetPasswordToken}`;
 
-        // Send the password reset link to the user's email
-        const mailSenderResponse = await mailSender(
+        // Send the password reset link via email
+        const mailResponse = await mailSender(
             email,
             "Password Reset Link",
             passwordResetLinkTemplate(resetPasswordUrl)
         );
 
-        if (!mailSenderResponse) {
+        if (!mailResponse) {
             return res.status(500).json({
                 success: false,
-                message: "Error occurred while sending the reset password URL",
+                message: "Failed to send password reset email.",
             });
         }
-        // Respond to the user
+
         return res.status(200).json({
             success: true,
-            message:
-                "Password reset link has been successfully sent to your email.",
-            resetPasswordToken,
+            message: "Password reset link sent successfully.",
         });
     } catch (error) {
-        console.error("Error while sending reset password link:", error);
+        console.error("Error sending password reset link:", error);
         return res.status(500).json({
             success: false,
-            message:
-                "An unexpected error occurred while processing the password reset request",
-            error: error.message,
+            message: "An error occurred while processing your request.",
         });
     }
 };
 
-// Function to reset the user's password using the reset token
-exports.resetPassword = async (req, res) => {
+// Function to reset password using the reset token
+export const resetPassword = async (req, res) => {
     try {
         const { resetPasswordToken, newPassword, confirmPassword } = req.body;
 
-        // Ensure passwords match
+        // Validate passwords match
         if (newPassword !== confirmPassword) {
             return res.status(400).json({
                 success: false,
-                message: `Passwords do not match booyah.${
-                    (newPassword, confirmPassword)
-                } Please try again.`,
+                message: "Passwords do not match. Please try again.",
             });
         }
 
-        // Find user by reset token
+        // Find user with the matching reset token
         const user = await User.findOne({ resetPasswordToken });
         if (!user) {
             return res.status(404).json({
                 success: false,
-                message: "Invalid token. Please request a new password reset.",
+                message: "Invalid or expired reset token.",
             });
         }
 
-        // Check if the reset token has expired
-        if (user.resetPasswordTokenExpires < Date.now()) {
+        // Check if the token has expired
+        if (Date.now() > user.resetPasswordTokenExpires) {
             return res.status(400).json({
                 success: false,
-                message:
-                    "Token has expired. Please request a new password reset.",
+                message: "Token has expired. Please request a new reset link.",
             });
         }
 
-        // Hash the new password before saving it
-        const hashedPassword = await bcrypt.hash(newPassword, 10);
-
-        // Update the user's password
-        user.password = hashedPassword;
-        // Clear the reset token and expiration time after successful reset
+        // Hash the new password and update user details
+        user.password = await bcrypt.hash(newPassword, 10);
         user.resetPasswordToken = undefined;
         user.resetPasswordTokenExpires = undefined;
-
-        // Save the updated user
         await user.save();
 
-        // Respond with success
         return res.status(200).json({
             success: true,
-            message: "Password successfully reset.",
+            message: "Password reset successful.",
         });
     } catch (error) {
-        console.error("Error while resetting password:", error);
+        console.error("Error resetting password:", error);
         return res.status(500).json({
             success: false,
-            message:
-                "An unexpected error occurred while processing the password reset request.",
-            error: error.message,
+            message: "An error occurred while resetting your password.",
         });
     }
 };
 
-// Change password function (for users who remember their old password)
-exports.changePassword = async (req, res) => {
+// Function to allow an authenticated user to change their password
+export const changePassword = async (req, res) => {
     try {
         const { oldPassword, newPassword, confirmPassword } = req.body;
         const userId = req.user.id;
-        // Check for missing fields
+
+        // Ensure all required fields are provided
         if (!oldPassword || !newPassword || !confirmPassword) {
             return res.status(400).json({
                 success: false,
-                message:
-                    "All fields (oldPassword, newPassword, confirmPassword) are required.",
+                message: "All fields are required.",
             });
         }
 
-        // Ensure new password matches confirmation password
+        // Validate new passwords match
         if (newPassword !== confirmPassword) {
             return res.status(400).json({
                 success: false,
-                message: "New password and confirmation password do not match.",
+                message: "New passwords do not match.",
             });
         }
 
-        // Find the user by email
+        // Fetch user from the database
         const user = await User.findById(userId);
         if (!user) {
             return res.status(404).json({
@@ -169,42 +153,39 @@ exports.changePassword = async (req, res) => {
                 message: "User not found.",
             });
         }
-        // Check if the old password matches the stored password
-        const isOldPasswordCorrect = await bcrypt.compare(
+
+        // Verify old password before allowing the change
+        const isPasswordCorrect = await bcrypt.compare(
             oldPassword,
             user.password
         );
-        if (!isOldPasswordCorrect) {
+        if (!isPasswordCorrect) {
             return res.status(401).json({
                 success: false,
-                message: "The old password is incorrect.",
+                message: "Incorrect old password.",
             });
         }
-        // Hash the new password
-        const hashedPassword = await bcrypt.hash(newPassword, 10);
 
-        // Update the user's password in the database
-        user.password = hashedPassword;
+        // Hash the new password and update the user document
+        user.password = await bcrypt.hash(newPassword, 10);
         await user.save();
 
-        // Send a confirmation email
+        // Send confirmation email about the password change
         await mailSender(
             user.email,
-            "Security Alert: Your Password Was Changed ",
+            "Your Password Was Changed",
             passwordUpdatedTemplate(user.email, user.firstName)
         );
 
-        // Respond with success
         return res.status(200).json({
             success: true,
-            message: "Password has been updated successfully.",
+            message: "Password updated successfully.",
         });
     } catch (error) {
-        console.error("Error while changing password:", error);
+        console.error("Error changing password:", error);
         return res.status(500).json({
             success: false,
             message: "An error occurred while updating the password.",
-            error: error.message,
         });
     }
 };
