@@ -79,7 +79,7 @@ const passwordResetLinkTemplate = (url) => {
             <a href="https://studynotion-edtech-project.vercel.app">
                 <img class="logo" src="https://res.cloudinary.com/dqhsaqbys/image/upload/v1741016843/studyNotion/logos/ijtgffa1jns9dhxbm7b1.png" alt="StudyNotion Logo">
             </a>
-            <div class="message">Password Reset Request booyahhh</div>
+            <div class="message">Password Reset Request</div>
             <div class="body">
                 <p>Dear User,</p>
                 <p>We received a request to reset your password. To reset your password, please click the link below:</p>
@@ -95,4 +95,4 @@ const passwordResetLinkTemplate = (url) => {
     
     </html>`;
 };
-module.exports = passwordResetLinkTemplate;
+export default passwordResetLinkTemplate;
