@@ -1,22 +1,29 @@
-const mongoose = require("mongoose");
+import mongoose from "mongoose";
 
 // Define the Section schema
-const sectionSchema = new mongoose.Schema({
-    sectionName: {
-        type: String, 
-        required: true, 
-        trim: true, 
-    },
-    subSection: [
-        {
-            type: mongoose.Schema.Types.ObjectId,
+const sectionSchema = new mongoose.Schema(
+    {
+        // Name of the section
+        sectionName: {
+            type: String,
             required: true,
-            ref: "SubSection",
+            trim: true, // Removes leading/trailing spaces
         },
-    ],
-}, {
-    timestamps: true, // Automatically adds createdAt and updatedAt fields
-});
+
+        // Array of subsections associated with this section
+        subSection: [
+            {
+                type: mongoose.Schema.Types.ObjectId,
+                required: true,
+                ref: "SubSection", // References the SubSection model
+            },
+        ],
+    },
+    {
+        timestamps: true, // Automatically adds `createdAt` and `updatedAt`
+    }
+);
 
 // Create and export the Section model
-module.exports = mongoose.model("Section", sectionSchema);
+const Section = mongoose.model("Section", sectionSchema);
+export default Section;
