@@ -1,7 +1,7 @@
-const Razorpay = require("razorpay");
+import Razorpay from "razorpay";
 
 // Initialize Razorpay instance using environment variables
-exports.instance = new Razorpay({
-    key_id: process.env.RAZORPAY_KEY_ID,    // Razorpay API Key
+export const instance = new Razorpay({
+    key_id: process.env.RAZORPAY_KEY_ID, // Razorpay API Key
     key_secret: process.env.RAZORPAY_SECRET_KEY, // Razorpay Secret Key
 });
