@@ -2,13 +2,18 @@
 //
 // 1. createTag
 // 2. showAllTags
-const Tag = require("../models/Tag");
 
-exports.createTag = async (req, res) => {
+// Import the Tag model
+import Tag from "../models/Tag.js";
+
+/**
+ * Controller to create a new tag.
+ */
+export const createTag = async (req, res) => {
     try {
         const { tagName } = req.body;
 
-        // Validate request body to ensure the 'name' field is provided
+        // Validate request body to ensure 'tagName' is provided
         if (!tagName) {
             return res.status(400).json({
                 success: false,
@@ -16,8 +21,8 @@ exports.createTag = async (req, res) => {
             });
         }
 
-        // Check if the tag already exists
-        const existingTag = await Tag.findOne({ tagName: tagName });
+        // Check if the tag already exists in the database
+        const existingTag = await Tag.findOne({ tagName });
         if (existingTag) {
             return res.status(409).json({
                 success: false,
@@ -25,16 +30,15 @@ exports.createTag = async (req, res) => {
             });
         }
 
-        // Create tag in the database with the provided details
-        const tagDetails = await Tag.create({
-            tagName: tagName,
-        });
+        // Create a new tag in the database
+        const tagDetails = await Tag.create({ tagName });
 
         console.log("Tag created successfully:", tagDetails);
 
         return res.status(201).json({
             success: true,
             message: "Tag created successfully.",
+            tag: tagDetails, // Returning the created tag for confirmation
         });
     } catch (error) {
         console.error(`Error creating tag: ${error.message}`);
@@ -42,14 +46,20 @@ exports.createTag = async (req, res) => {
             success: false,
             message:
                 "An unexpected error occurred while creating the tag. Please try again later.",
+            error:
+                process.env.NODE_ENV === "development"
+                    ? error.message
+                    : "Internal Server Error",
         });
     }
 };
 
-// Controller to show all Tags
-exports.showAllTags = async (req, res) => {
+/**
+ * Controller to retrieve all tags.
+ */
+export const showAllTags = async (req, res) => {
     try {
-        // Fetch all tags, selecting only their name and description fields
+        // Fetch all tags, selecting only the 'tagName' field
         const allTags = await Tag.find({}, "tagName");
 
         return res.status(200).json({
@@ -62,6 +72,10 @@ exports.showAllTags = async (req, res) => {
             success: false,
             message:
                 "An unexpected error occurred while retrieving tags. Please try again later.",
+            error:
+                process.env.NODE_ENV === "development"
+                    ? error.message
+                    : "Internal Server Error",
         });
     }
 };
